Introduce "modals" directory and move modals there
