+++ packages/scrabble-solver/src/components/index.ts
@@ -6,10 +6,8 @@ export { default as Dictionary } from './Dictionary';
 export { default as DictionaryInput } from './DictionaryInput';
 export { default as EmptyState } from './EmptyState';
 export { default as Key } from './Key';
-export { default as KeyMap } from './KeyMap';
 export { default as Loading } from './Loading';
 export { default as Logo } from './Logo';
-export { default as Menu } from './Menu';
 export { default as Modal } from './Modal';
 export { default as NavButtons } from './NavButtons';
 export { default as NotFound } from './NotFound';
@@ -17,11 +15,9 @@ export { default as PlainTiles } from './PlainTiles';
 export { default as Progress } from './Progress';
 export { default as Rack } from './Rack';
 export { default as Radio } from './Radio';
-export { default as RemainingTiles } from './RemainingTiles';
 export { default as Results } from './Results';
 export { default as ResultsInput } from './ResultsInput';
 export { default as Screen } from './Screen';
-export { default as Settings } from './Settings';
 export { default as Solver } from './Solver';
 export { default as SolverMobile } from './SolverMobile';
 export { default as Splash } from './Splash';
@@ -31,4 +27,3 @@ export { default as SvgFontFix } from './SvgFontFix';
 export { default as Tile } from './Tile';
 export { useTooltip } from './Tooltip';
 export { default as Well } from './Well';
-export { default as Words } from './Words';
