+++ packages/scrabble-solver/src/modals/KeyMap/components/Mapping/Mapping.module.scss
