+++ packages/scrabble-solver/src/modals/KeyMap/components/Mapping/index.ts
