+++ packages/scrabble-solver/src/modals/RemainingTiles/components/Character/Character.module.scss
