+++ packages/scrabble-solver/src/modals/RemainingTiles/components/Character/Character.tsx
@@ -2,13 +2,11 @@ import { BLANK } from '@scrabble-solver/constants';
 import classNames from 'classnames';
 import { FunctionComponent } from 'react';
 
+import { Progress, Tile } from 'components';
 import { LOCALE_FEATURES } from 'i18n';
 import { REMAINING_TILES_TILE_SIZE } from 'parameters';
 import { selectCharacterPoints, selectLocale, useTypedSelector } from 'state';
 import { RemainingTile } from 'types';
-
-import Progress from '../Progress';
-import Tile from '../Tile';
 
 import styles from './Character.module.scss';
