+++ packages/scrabble-solver/src/modals/RemainingTiles/components/Character/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Character';
