+++ packages/scrabble-solver/src/modals/RemainingTiles/components/index.ts
@@ -1,0 +1,1 @@
+export { default as Character } from './Character';
