+++ packages/scrabble-solver/src/modals/Settings/components/AutoGroupTilesSetting/AutoGroupTilesSetting.module.scss
