+++ packages/scrabble-solver/src/modals/Settings/components/AutoGroupTilesSetting/index.ts
