+++ packages/scrabble-solver/src/modals/Settings/components/AutoGroupTilesSetting/lib.ts
