+++ packages/scrabble-solver/src/modals/Settings/components/ConfigSetting/ConfigSetting.module.scss
