+++ packages/scrabble-solver/src/modals/Settings/components/ConfigSetting/index.ts
