+++ packages/scrabble-solver/src/modals/Settings/components/ConfigSetting/options.ts
