+++ packages/scrabble-solver/src/modals/Settings/components/ConfigSetting/types.ts
