+++ packages/scrabble-solver/src/modals/Settings/components/LocaleSetting/LocaleSetting.module.scss
