+++ packages/scrabble-solver/src/modals/Settings/components/LocaleSetting/LocaleSetting.tsx
@@ -3,9 +3,8 @@ import classNames from 'classnames';
 import { ChangeEvent, FunctionComponent } from 'react';
 import { useDispatch } from 'react-redux';
 
+import { Radio } from 'components';
 import { selectLocale, settingsSlice, useTypedSelector } from 'state';
-
-import Radio from '../../../Radio';
 
 import styles from './LocaleSetting.module.scss';
 import options from './options';
