+++ packages/scrabble-solver/src/modals/Settings/components/LocaleSetting/index.ts
