+++ packages/scrabble-solver/src/modals/Settings/components/LocaleSetting/options.ts
