+++ packages/scrabble-solver/src/modals/Settings/components/LocaleSetting/types.ts
