+++ packages/scrabble-solver/src/modals/Settings/components/index.ts
