+++ packages/scrabble-solver/src/modals/Settings/index.ts
