+++ packages/scrabble-solver/src/modals/index.ts
@@ -1,0 +1,5 @@
+export { default as KeyMap } from './KeyMap';
+export { default as Menu } from './Menu';
+export { default as RemainingTiles } from './RemainingTiles';
+export { default as Settings } from './Settings';
+export { default as Words } from './Words';
